+++ cypress/e2e/types.objects.cy.ts
@@ -1,5 +1,5 @@
 describe('Types page flow', () => {
-  const name = `e2e-test-data-${Date.now()}`
+  const name = `e2e-test-type-object-${Date.now()}`
   const fileName = `${name}.yaml`
   const thingsToDelete: string[] = []
 
@@ -14,21 +14,38 @@ describe('Types page flow', () => {
     cy.get('[data-test="new-type-create-btn"]').click()
     cy.wait(500)
     cy.url().should('include', `/types/${name}`)
-
-    // test description has placeholder text "Click to add description" and type chip picker with "void" value
-    // change type to object
-    // enter a description
-    // verify No Properties message is visible
-    // add 3 properties to the object
+    cy.get('[data-test="root-description-placeholder"]').should('be.visible').and('contain', 'Click to add description')
+    cy.get('[data-test="root-type-chip-picker"] [data-test="type-chip"]').should('be.visible').and('contain', 'void')    
+    cy.get('[data-test="root-type-chip-picker"] [data-test="type-chip"]').click()
+    cy.get('[data-test="built-in-type-object"]').click()
   })
 
   // Clean up any documents created during tests
   afterEach(() => {
     thingsToDelete.forEach((thing) => {
       cy.request({
-        method: 'DELETE',
+        method: 'PUT',    
         url: thing,
+        headers: {"Content-Type": "application/json"},
+        body: {"_locked": false, "root":{"name":""}},
         failOnStatusCode: false
+      }).then((response) => {
+        if (response.status === 200) {
+          cy.log(`Successfully unlocked ${thing}`)
+          cy.request({
+            method: 'DELETE',
+            url: thing,
+            failOnStatusCode: false
+          }).then((response) => {
+            if (response.status === 200) {
+              cy.log(`Successfully deleted ${thing}`)
+            } else {
+              cy.log(`Failed to delete ${thing}: ${response.status}`)
+            }
+          })
+        } else {
+          cy.log(`Failed to unlock ${thing}: ${response.status}`)
+        }
       })
     })
     cy.visit('/types')
